refactor: unify code structure by renaming and moving file, switch to createCredentialsClient
